Pave the way for astral character support.
No functional change, but redefine the meaning of the old "size"
member of the rufl_character_set structure to allow for the
addition of extension structures in future. This change is
backwards compatible as it is reusing previously unused bits in
the size field (which will be set to zero in all existing
RUfl_caches). Rename the "size" field to "metadata" which better
reflects its new usage.

Update rufl_character_set_test and rufl_dump_state to follow this
change (and fix up their parameter types while we're here).