Optimise substitution table storage
Consider each Unicode plane independently (as they have very
different properties). This means building a table for each
plane and allows us to reduce the size of each entry in the
CHD-addressed table from 64 to 32bits (which provides a
significant immediate saving).

Also introduce a direct linear mapping backend. This stores the
table in a series of 256-entry blocks which are addressed from
a fixed-size index. Block entries are either 8 or 16 bits wide
(depending upon the number of fonts found on the system). This
restores some of the storage efficiency of the old "giant array"
approach, which is generally more efficient than a CHD (or other)
hash-based implementation where the load factor is reasonably high
(or the glyph:block ratio is sufficiently high).

Select the direct or CHD storage mechanism based upon an estimate
of the storage size for the data collected for a plane. In the
testing I have performed (with the same fonts available as before)
the combined effect of the above is to reduce the storage used
significantly.

Without the 8bit direct mapping entry size (which is a somewhat
unfair comparison because even the "giant array" didn't have that
feature) we see:

Plane  Codepoints  Blocks  Backend  Storage  Alternative

  1    51483       224     Direct   115480   311328 (CHD)
  2     1981        13     Direct     7448     9760 (CHD)
  3     2293       201     CHD       17952   103704 (Direct)

Total  55757                        140880 (~= 2.5 bytes/glyph)

The other 14 planes have no glyph coverage at all, so require no
storage.

With the 8bit direct mapping, we see:

Plane  Codepoints  Blocks  Backend  Storage  Alternative

  1    51483       224     Direct    57880   311328 (CHD)
  2     1981        13     Direct     3864     9760 (CHD)
  3     2293       201     CHD       17952   103704 (Direct)

Total  55757                         79696 (~= 1.4 bytes/glyph)

In summary:

  * separating the planes has shaved ~50% off the storage required
    by the CHD backend
  * introducing the direct mapping backend has shaved a further
    ~60% off that
  * using 8bit direct mapping has shaved another ~50% off that

Cumulatively, then, storage requirements are now ~86% smaller
than with CHD only (and about 40% less than the BMP-only
"giant table", but now with astral character support).