Partially revert public API type changes
a4c41198 made a variety of consistency changes to the public API,
including changing the type of the "string" parameter passed to
many entry points from const char * to const uint8_t *, as that
better reflects the data. However, this then forces the user of
the API to explicitly cast when passing string constants, or
other strings (which, would be passed to standard library APIs
as const char *, even if UTF-8 encoded).

Revert this part of the change so the type of "string" is once
more const char * and cast to the type we actually want internally.